@@ -1,10 +1,9 @@
 package com.smk.modul1.category.service;
 
+import java.util.List;
+
 import com.smk.modul1.dto.input.CategoryInput;
 import com.smk.modul1.dto.output.CategoryOutput;
-import com.smk.modul1.model.Category;
-
-import java.util.List;
 
 public interface CategoryService {
     CategoryOutput addOne(CategoryInput input);
