@@ -1,6 +1,5 @@
 package com.smk.modul1.base;
 
-import com.smk.modul1.base.BaseResponse;
 import org.springframework.http.HttpStatus;
 import org.springframework.validation.BindingResult;
 import org.springframework.validation.FieldError;
@@ -13,22 +12,20 @@ public class BaseController {
     @ExceptionHandler(MethodArgumentNotValidException.class)
     @ResponseStatus(HttpStatus.BAD_REQUEST)
     @ResponseBody
-    public BaseResponse<?>
-    handleValidationError(MethodArgumentNotValidException ex) {
+    public BaseResponse<?> handleValidationError(MethodArgumentNotValidException ex) {
         BindingResult bindingResult = ex.getBindingResult();
         FieldError fieldError = bindingResult.getFieldError();
         assert fieldError != null;
         String defaultMessage = fieldError.getDefaultMessage();
-        return new BaseResponse<>(HttpStatus.BAD_REQUEST.value(),false,
+        return new BaseResponse<>(HttpStatus.BAD_REQUEST.value(), false,
                 defaultMessage);
     }
 
     @ExceptionHandler(Exception.class)
     @ResponseStatus(HttpStatus.INTERNAL_SERVER_ERROR)
     @ResponseBody
-    public BaseResponse<?>
-    handleException(Exception ex) {
-        return new BaseResponse<>(HttpStatus.INTERNAL_SERVER_ERROR.value(),false,
+    public BaseResponse<?> handleException(Exception ex) {
+        return new BaseResponse<>(HttpStatus.INTERNAL_SERVER_ERROR.value(), false,
                 ex.getMessage());
     }
 }
