@@ -1,7 +1,10 @@
 package com.smk.modul1.product.repository;
 
-import com.smk.modul1.model.Product;
 import org.springframework.data.repository.CrudRepository;
+import org.springframework.stereotype.Repository;
 
+import com.smk.modul1.model.Product;
+
+@Repository
 public interface ProductRepository extends CrudRepository<Product, Long> {
 }
